% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_evidence}
\alias{calculate_evidence}
\title{Calculates data evidence matrix E from distances matrix D.}
\usage{
calculate_evidence(D)
}
\arguments{
\item{D}{Distances matrix of size N x c.}
}
\value{
Matrix of size N x c.
}
\description{
Calculates data evidence matrix E from distances matrix D.
}
