% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{compute_distance}
\alias{compute_distance}
\title{Compute a distance matrix}
\usage{
compute_distance(ctx, distance = "cdist", control = distance_control())
}
\arguments{
\item{ctx}{A named list representing the model context.}

\item{distance}{Either a built-in distance name or a custom function.}

\item{control}{Control parameters for built-in distance methods.}
}
\value{
A distance matrix.
}
\description{
Computes the raw distance matrix using either one of the built-in
implementations or a user-defined distance function.
}
\details{
Built-in implementations:
\itemize{
\item "cdist"
\item "adaptive"
}

User-defined distance functions must accept a single argument \code{ctx},
a named list containing the variables available in the current model.

Note that \code{rdist::cdist} returns \eqn{D = [d_{jk}]}, and hence its
result must be raised to the power of \eqn{2}, whereas \code{adist}
return as \eqn{D = [d^2_{jk}]}.
}
