% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{distance_control}
\alias{distance_control}
\title{Control parameters for built-in distance functions}
\usage{
distance_control(rho = NULL)
}
\arguments{
\item{rho}{Optional tuning parameter for adaptive distance.}
}
\value{
A list of control parameters.
}
\description{
Control parameters for built-in distance functions
}
