% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimate_super_T}
\alias{estimate_super_T}
\title{Estimated T matrix with typicalities in semi-supervised case.}
\usage{
estimate_super_T(D, superF, alpha, gammas, b = 1)
}
\arguments{
\item{D}{Distances matrix of size N x c.}

\item{superF}{Binary supervision matrix of size N x c.}

\item{alpha}{Scaling factor, a floating point > 0 regulating the impact of partial supervision.}

\item{gammas}{a c-vector of cluster-specific gamma hyperparameters.}

\item{b}{a scalar weighting the contribution of possibilistic membership in
SPFCM (semi-supervised possibilistic fuzzy c-means) model.
It is set to 1 by default for other semi-supervised models.}
}
\description{
Estimated T matrix with typicalities in semi-supervised case.
}
