% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimate_U}
\alias{estimate_U}
\title{Estimated U matrix with memberships in semi-supervised case.}
\usage{
estimate_U(D, superF, alpha)
}
\arguments{
\item{D}{Distances matrix of size N x c.}

\item{superF}{Binary supervision matrix of size N x c.}

\item{alpha}{Scaling factor, a floating point > 0 regulating the impact of partial supervision.}
}
\description{
Estimated U matrix with memberships in semi-supervised case.
}
