% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcm.R
\name{FCM}
\alias{FCM}
\title{Fuzzy C-Means model.}
\usage{
FCM(
  X,
  C,
  U = NULL,
  max_iter = 200,
  conv_criterion = 1e-04,
  function_dist = rdist::cdist
)
}
\arguments{
\item{X}{a matrix \emph{X} with predictor variables.}

\item{C}{a number of clusters to find.}

\item{U}{optionally: a first memberships matrix to initialize the algorithm.
Used mainly for reproducibility to compare calculations with other packages
(e.g. in Python).}

\item{function_dist}{A function of two arguments: matrices X and V of the same
number of columns.
It should return a matrix of (nrow(X) x nrow(V)) of distances
between each row of X and all rows of V.
In case of Euclidean distance, the result should not be squared!}
}
\description{
Unsupervised Fuzzy c-Means algorithm.
}
