% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gamma_fcm}
\alias{gamma_fcm}
\title{Aggregates elements of DHE and DVE matrices in a step to build
evidence matrix E.}
\usage{
gamma_fcm(dhe, dve)
}
\arguments{
\item{dhe}{DHE matrix of size Nc x c.}

\item{dve}{DVE matrix of size Nc x c.}
}
\value{
Matrix of size Nc x 1.
}
\description{
Aggregates elements of DHE and DVE matrices in a step to build
evidence matrix E.
}
