% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{init_gamma}
\alias{init_gamma}
\title{Initialization procedure to calculate values of gamma hyperparameters.}
\usage{
init_gamma(.model, .X)
}
\arguments{
\item{.model}{estimated model of class \code{fcm}}

\item{.X}{features matrix of size N x c}
}
\description{
Initialization procedure to calculate values of gamma hyperparameters.
}
