% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcm.R
\name{PCM}
\alias{PCM}
\title{Possibilistic c-Means model.}
\usage{
PCM(
  X,
  C,
  U = NULL,
  gammas = NULL,
  initFCM = NULL,
  max_iter = 200,
  conv_criterion = 1e-04,
  function_dist = rdist::cdist
)
}
\arguments{
\item{X}{Features matrix \emph{X}.}

\item{C}{Number of clusters.}

\item{U}{Optionally: a concrete initialization memberships matrix.
Used mainly for reproducibility.
Default value \code{NULL} - algorithm uses random initialization in such case.}

\item{gammas}{Optionally: a vector of cluster-specific gamma hyperparameters.
Default value: \code{NULL}. In such case, the initialization value depends on
\code{initFCM} value. If \code{initFCM} is \code{NULL}, then vector filled with ones is used.
If \code{initFCM} is not \code{NULL}, then Fuzzy c-Means model is fitted, and the algorithm
implemented in \code{init_gamma} function is used to calculate the cluster-specific
gamma hyperparameters.}

\item{initFCM}{Default value: \code{NULL}. If \code{gammas} is not \code{NULL}, and \code{initFCM} is not \code{NULL},
then special algorithm based on creating Fuzzy c-Means model and appropriately
calculating values of hyperparameters gamma is initiated.}

\item{max_iter}{Maximum number of iterations. Default value: 200.}

\item{conv_criterion}{Convergence criterion value used at the end of each iteration of
Alternating Optimization algorithm.}

\item{function_dist}{Optionally: a function of two arguments: matrices \emph{X} and \emph{V} of the same
number of columns.
It should return a matrix of (nrow(X) x nrow(V)) of distances
between each row of \emph{X} and all rows of \emph{V}.
In case of Euclidean distance, the result should not be squared!}
}
\value{
An object of class \code{pcm} containing:
\describe{
\item{U}{An \eqn{N \times c} matrix of cluster memberships.}
\item{V}{A \eqn{c \times p} matrix of cluster prototypes.}
\item{function_dist}{An object of class \code{function} used to calculate distances.}
\item{counter}{Integer number of iterations until convergence.}
\item{V_history}{A list of length \code{counter} with \eqn{c \times p}
prototypes matrices estimated in each loop of the algorithm.}
\item{U_history}{A list of length \code{counter} with \eqn{N \times c}
memberships matrices estimated in each loop of the algorithm.}
\item{Phi_history}{A list of length \code{counter} with \eqn{N \times c}
phi weights in each loop of the algorithm.}
}
}
\description{
Unsupervised Possibilistic c-Means algorithm.
}
\examples{
X <- matrix(rnorm(99), ncol = 3)
model_pcm <- PCM(X = X, C = 3)
print(model_pcm$V)

}
