% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssfcm.R
\name{predict.ssfcm}
\alias{predict.ssfcm}
\title{Predict method for SSFCM objects}
\usage{
\method{predict}{ssfcm}(object, .X, ...)
}
\arguments{
\item{object}{An object of class \code{ssfcm}}

\item{...}{Not used}

\item{X}{New data matrix}
}
\value{
A data frame with predicted labels
}
\description{
Predict method for SSFCM objects
}
