% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spfcm.R
\name{SPFCM}
\alias{SPFCM}
\title{Semi-Supervised Possibilistic Fuzzy C-Means clustering model}
\usage{
SPFCM(
  X,
  C,
  U = NULL,
  Tp = NULL,
  gammas = NULL,
  initFCM = NULL,
  max_iter = 200,
  conv_criterion = 1e-04,
  function_dist = rdist::cdist,
  store_history = FALSE,
  alpha = NULL,
  superF = NULL,
  a = 1,
  b = 1
)
}
\arguments{
\item{X}{A numeric feature matrix.}

\item{C}{Integer specifying the number of clusters.}

\item{U}{Optional initial membership matrix.
Primarily intended for reproducibility purposes.
If \code{NULL} (default), the algorithm uses a random initialization.}

\item{Tp}{DESCRIPTIO TO BE DONE.}

\item{gammas}{Optional vector of cluster-specific gamma hyperparameters.
If \code{NULL} (default), the initialization strategy depends on the value
of \code{initFCM}.

If \code{initFCM} is \code{NULL}, a vector of ones is used.
Otherwise, a Fuzzy C-Means model is first fitted, and the
\code{init_gamma()} function is used to estimate the cluster-specific
gamma hyperparameters.}

\item{initFCM}{Optional fitted Fuzzy C-Means model used to initialize
cluster-specific gamma hyperparameters via weighted averaging.
If \code{NULL} (default), no preliminary Fuzzy C-Means initialization is used.
If provided, this argument is effective only when \code{gammas} is \code{NULL}.}

\item{max_iter}{Maximum number of iterations.
Defaults to \code{200}.}

\item{conv_criterion}{Convergence threshold used at the end of each
iteration of the Alternating Optimization algorithm.}

\item{function_dist}{Optional distance function.
The function must accept two matrices, \code{X} and \code{V}, with the same
number of columns, and return a matrix of size
\verb{nrow(X) x nrow(V)} containing distances between each row of \code{X}
and each row of \code{V}.

For the Euclidean distance, the returned distances should not be squared.
Defaults to \code{\link[rdist:rdist]{rdist::cdist()}}.}

\item{store_history}{Logical indicating whether optimization
histories should be stored. If \code{FALSE}, the returned object
will contain \code{NULL} history fields. Defaults to \code{TRUE}.}

\item{alpha}{Positive scaling factor regulating the impact of
partial supervision.}

\item{superF}{Binary supervision matrix of the same dimensions as \code{U},
indicating the available partial supervision information.}

\item{a}{DESCRIPTION TO BE DONE.}

\item{b}{DESCRIPTION TO BE DONE.}
}
\value{
An object of class \code{spfcm} containing:
\describe{
\item{U}{An \eqn{N \times C} typicalities matrix.}
\item{V}{A \eqn{C \times p} matrix of cluster prototypes.}
\item{function_dist}{The distance function used by the model.}
\item{counter}{Number of iterations performed until convergence.}
\item{gammas}{Vector of cluster-specific gamma hyperparameters.}
\item{alpha}{Value of scaling factor.}
\item{U_history}{If \code{store_history = TRUE}, a list of length
\code{counter} containing membership matrices estimated at each
iteration; otherwise \code{NULL}.}
\item{V_history}{If \code{store_history = TRUE}, a list of length
\code{counter} containing prototype matrices estimated at each
iteration; otherwise \code{NULL}.}
\item{Phi_history}{If \code{store_history = TRUE}, a list of length
\code{counter} containing phi-weight matrices estimated at each
iteration; otherwise \code{NULL}.}
}
}
\description{
Fits a Semi-Supervised Possibilistic Fuzzy C-Means (SPFCM) clustering model using the Alternating
Optimization algorithm.
}
\examples{
X <- matrix(rnorm(100), ncol = 2)

superF <- matrix(0, nrow = nrow(X), ncol = 2)

superF[1:10, 1] <- 1
superF[11:20, 2] <- 1

model_spfcm <- SPFCM(
  X = X,
  C = 2,
  superF = superF,
  alpha = 1
)

print(model_spfcm$V)

}
\references{
V. Antoine, J.A. Guerrero, & G. Romero (2022).
\emph{Possibilistic fuzzy c-means with partial supervision}.
Fuzzy Sets and Systems (449), 162-186.
https://doi.org/10.1016/j.fss.2022.08.003
}
