% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssfcm.R
\name{SSFCM}
\alias{SSFCM}
\title{Semi-Supervised Fuzzy C-Means model.}
\usage{
SSFCM(
  X,
  C,
  U = NULL,
  max_iter = 200,
  conv_criterion = 1e-04,
  function_dist = rdist::cdist,
  alpha = NULL,
  superF = NULL
)
}
\arguments{
\item{X}{a matrix \emph{X} with predictor variables.}

\item{C}{a number of clusters to find.}

\item{U}{optionally: a first memberships matrix to initialize the algorithm.
Used mainly for reproducibility to compare calculations with other packages
(e.g. in Python).}

\item{function_dist}{A function of two arguments: matrices X and V of the same
number of columns.
It should return a matrix of (nrow(X) x nrow(V)) of distances
between each row of X and all rows of V.
In case of Euclidean distance, the result should not be squared!}

\item{alpha}{the scaling factor, a floating point > 0.}

\item{F_}{the supervision  binary matrix of the same dimension as \emph{U}.}
}
\description{
If \emph{alpha} and \emph{F_} are not supplied (their default values are \code{NULL}),
then a regular unsupervised Fuzzy C-Means algorithm is fitted.
}
