\name{U_underimpact}
\alias{U_underimpact}
\docType{data}
\title{Initialization matrix to analyze underimpact in iris data.}
\description{
  This dataset provides a concrete initialization of membership matrix
  specific to the iris data that exhibits the phenomenon of underimpact
  of partial supervision in semi-supervised fuzzy clustering.
}
\usage{data(U_underimpact)}
\format{A matrix of size 150 x 3.}
\keyword{datasets}