% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimate_T}
\alias{estimate_T}
\title{Estimated T matrix with typicalities in unsupervised case.}
\usage{
estimate_T(D, gammas)
}
\arguments{
\item{D}{Distances matrix of size N x c.}

\item{gammas}{a c-vector of cluster-specific gamma hyperparameters.}
}
\description{
Estimated T matrix with typicalities in unsupervised case.
}
