% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcm.R
\name{FCM}
\alias{FCM}
\title{Fuzzy C-Means model.}
\usage{
FCM(
  X,
  C,
  U = NULL,
  max_iter = 200,
  conv_criterion = 1e-04,
  function_dist = rdist::cdist
)
}
\arguments{
\item{X}{Features matrix \emph{X}.}

\item{C}{Number of clusters.}

\item{U}{Optionally: a concrete initialization memberships matrix.
Used mainly for reproducibility.
Default value \code{NULL} - algorithm uses random initialization in such case.}

\item{max_iter}{Maximum number of iterations. Default value: 200.}

\item{conv_criterion}{Convergence criterion value used at the end of each iteration of
Alternating Optimization algorithm.}

\item{function_dist}{Optionally: a function of two arguments: matrices \emph{X} and \emph{V} of the same
number of columns.
It should return a matrix of (nrow(X) x nrow(V)) of distances
between each row of \emph{X} and all rows of \emph{V}.
In case of Euclidean distance, the result should not be squared!}
}
\value{
An object of class \code{fcm} containing:
\describe{
\item{U}{An \eqn{N \times c} matrix of cluster memberships.}
\item{V}{A \eqn{c \times p} matrix of cluster prototypes.}
\item{function_dist}{An object of class \code{function} used to calculate distances.}
\item{counter}{Integer number of iterations until convergence.}
\item{V_history}{A list of length \code{counter} with \eqn{c \times p}
prototypes matrices estimated in each loop of the algorithm.}
\item{U_history}{A list of length \code{counter} with \eqn{N \times c}
memberships matrices estimated in each loop of the algorithm.}
\item{Phi_history}{A list of length \code{counter} with \eqn{N \times c}
phi weights in each loop of the algorithm.}
}
}
\description{
Unsupervised Fuzzy c-Means algorithm.
}
\examples{
X <- matrix(rnorm(100), ncol = 2)
model_fcm <- fussclust::FCM(X = X, C = 2)
print(model_fcm$V)

}
