% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcm.R
\name{PCM}
\alias{PCM}
\title{Possibilistic C-Means clustering model}
\usage{
PCM(
  X,
  C,
  U = NULL,
  gammas = NULL,
  initFCM = NULL,
  max_iter = 200,
  conv_criterion = 0.0001,
  function_dist = rdist::cdist
)
}
\arguments{
\item{X}{A numeric feature matrix.}

\item{C}{Integer specifying the number of clusters.}

\item{U}{Optional initial membership matrix.
Primarily intended for reproducibility purposes.
If \code{NULL} (default), the algorithm uses a random initialization.}

\item{gammas}{Optional vector of cluster-specific gamma hyperparameters.
If \code{NULL} (default), the initialization strategy depends on the value
of \code{initFCM}.

If \code{initFCM} is \code{NULL}, a vector of ones is used.
Otherwise, a Fuzzy C-Means model is first fitted, and the
\code{init_gamma()} function is used to estimate the cluster-specific
gamma hyperparameters.}

\item{initFCM}{Optional fitted Fuzzy C-Means model used to initialize
cluster-specific gamma hyperparameters via weighted averaging.
If \code{NULL} (default), no preliminary Fuzzy C-Means initialization is used.
If provided, this argument is effective only when \code{gammas} is \code{NULL}.}

\item{max_iter}{Maximum number of iterations.
Defaults to \code{200}.}

\item{conv_criterion}{Convergence threshold used at the end of each
iteration of the Alternating Optimization algorithm.}

\item{function_dist}{Optional distance function.
The function must accept two matrices, \code{X} and \code{V}, with the same
number of columns, and return a matrix of size
\verb{nrow(X) x nrow(V)} containing distances between each row of \code{X}
and each row of \code{V}.

For the Euclidean distance, the returned distances should not be squared.
Defaults to \code{\link[rdist:rdist]{rdist::cdist()}}.}
}
\value{
An object of class \code{pcm} containing:
\describe{
\item{U}{An \eqn{N \times C} membership matrix.}
\item{V}{A \eqn{C \times p} matrix of cluster prototypes.}
\item{function_dist}{The distance function used by the model.}
\item{counter}{Number of iterations performed until convergence.}
\item{gammas}{Vector of cluster-specific gamma hyperparameters.}
\item{V_history}{A list of length \code{counter} containing prototype
matrices estimated at each iteration.}
\item{U_history}{A list of length \code{counter} containing membership
matrices estimated at each iteration.}
\item{Phi_history}{A list of length \code{counter} containing phi-weight
matrices estimated at each iteration.}
}
}
\description{
Fits a Possibilistic C-Means (PCM) clustering model using the Alternating
Optimization algorithm.
}
\examples{
X <- matrix(rnorm(100), ncol = 2)

model_pcm <- fussclust::PCM(
  X = X,
  C = 2,
  initFCM = TRUE
)

print(model_pcm$V)

}
\references{
Krishnapuram, R., & Keller, J. (1993).
\emph{A possibilistic approach to clustering}.
IEEE Transactions on Fuzzy Systems, 1(2), 98–110.
https://doi.org/10.1109/91.227387
}
