% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sspcm.R
\name{predict.sspcm}
\alias{predict.sspcm}
\title{Soft assignment score function}
\usage{
\method{predict}{sspcm}(object, newdata)
}
\arguments{
\item{object}{object}

\item{newdata}{data to be predicted}
}
\value{
typicalities matrix
}
\description{
Soft assignment score function
}
