% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssfcm.R
\name{SSFCM}
\alias{SSFCM}
\title{Semi-Supervised Fuzzy C-Means clustering model}
\usage{
SSFCM(
  X,
  C,
  U = NULL,
  max_iter = 200,
  conv_criterion = 0.0001,
  function_dist = rdist::cdist,
  store_history = FALSE,
  alpha = NULL,
  superF = NULL
)
}
\arguments{
\item{X}{A numeric feature matrix.}

\item{C}{Integer specifying the number of clusters.}

\item{U}{Optional initial membership matrix.
Primarily intended for reproducibility purposes.
If \code{NULL} (default), the algorithm uses a random initialization.}

\item{max_iter}{Maximum number of iterations.
Defaults to \code{200}.}

\item{conv_criterion}{Convergence threshold used at the end of each
iteration of the Alternating Optimization algorithm.}

\item{function_dist}{Optional distance function.
The function must accept two matrices, \code{X} and \code{V}, with the same
number of columns, and return a matrix of size
\verb{nrow(X) x nrow(V)} containing distances between each row of \code{X}
and each row of \code{V}.

For the Euclidean distance, the returned distances should not be squared.
Defaults to \code{\link[rdist:rdist]{rdist::cdist()}}.}

\item{store_history}{Logical indicating whether optimization
histories should be stored. If \code{FALSE}, the returned object
will contain \code{NULL} history fields. Defaults to \code{TRUE}.}

\item{alpha}{Positive scaling factor regulating the impact of
partial supervision.}

\item{superF}{Binary supervision matrix of the same dimensions as \code{U},
indicating the available partial supervision information.}
}
\value{
An object of class \code{sspcm} containing:
\describe{
\item{U}{An \eqn{N \times C} typicalities matrix.}
\item{V}{A \eqn{C \times p} matrix of cluster prototypes.}
\item{function_dist}{The distance function used by the model.}
\item{counter}{Number of iterations performed until convergence.}
\item{alpha}{Value of scaling factor.}
\item{U_history}{If \code{store_history = TRUE}, a list of length
\code{counter} containing membership matrices estimated at each
iteration; otherwise \code{NULL}.}
\item{V_history}{If \code{store_history = TRUE}, a list of length
\code{counter} containing prototype matrices estimated at each
iteration; otherwise \code{NULL}.}
\item{Phi_history}{If \code{store_history = TRUE}, a list of length
\code{counter} containing phi-weight matrices estimated at each
iteration; otherwise \code{NULL}.}
}
}
\description{
Fits a Semi-Supervised Fuzzy C-Means (SSPCM) clustering model using the Alternating
Optimization algorithm.
}
\examples{
X <- matrix(rnorm(100), ncol = 2)

superF <- matrix(0, nrow = nrow(X), ncol = 2)

superF[1:10, 1] <- 1
superF[11:20, 2] <- 1

model_ssfcm <- SSFCM(
  X = X,
  C = 2,
  superF = superF,
  alpha = 1
)

print(model_ssfcm$V)

}
\references{
Kmita, K., Kaczmarek-Majer, K., & Hryniewicz, O. (2024).
\emph{Explainable Impact of Partial Supervision in Semi-Supervised
Fuzzy Clustering}.
IEEE Transactions on Fuzzy Systems, 1–10.
https://doi.org/10.1109/TFUZZ.2024.3370768
}
